requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
bcrypt==4.3.0
    # via paramiko
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
cryptography==44.0.2
    # via paramiko
flask==3.1.0
    # via repo-checker (pyproject.toml)
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via repo-checker (pyproject.toml)
gunicorn==23.0.0
    # via repo-checker (pyproject.toml)
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
packaging==24.2
    # via gunicorn
paramiko==3.5.1
    # via repo-checker (pyproject.toml)
pycparser==2.22
    # via cffi
pynacl==1.5.0
    # via paramiko
requests==2.32.3
    # via repo-checker (pyproject.toml)
smmap==5.0.2
    # via gitdb
urllib3==2.3.0
    # via requests
werkzeug==3.1.3
    # via flask
